requirements_test.txt
pytest
pytest-cov
pytest-mock
pytest_click
mock

-r requirements.txt
